88: Removed legacy transactions store
