+++ src/App.tsx
@@ -5,7 +5,7 @@ import { Web3ReactProvider } from '@web3-react/core';
 import Routes from './compositions/Routes/Routes';
 import { setLibrary } from './helpers/ethers';
 import useMapWeb3ReactToStore from './hooks/useMapWeb3ReactToStore';
-import { useTransactions } from './redux/stores/transactions-v2/transactionsHooks';
+import { useTransactions } from './redux/stores/transactions/transactionsHooks';
 import ToastsWidget from './widgets/ToastsWidget/ToastsWidget';
 
 const ConnectedApp = () => {
