--- src/redux/stores/transactions/Transactions.module.css
@@ -1,14 +1,0 @@
-.row {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.row > button {
-  margin-left: 4px;
-  margin-right: 8px;
-}
-
-.row:not(:last-child) {
-  margin-bottom: 16px;
-}
